@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "cp2k-input-tools"
-version = "0.8.2"
+version = "0.9.0"
 description = "Python tools to handle CP2K input files"
 authors = ["Tiziano Müller <tiziano.mueller@chem.uzh.ch>"]
 repository = "https://github.com/cp2k/cp2k-input-tools"
@@ -16,14 +16,13 @@ classifiers = [
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">=3.6.1,<4"
-transitions = "^0.7"
-pint = ">=0.15,<0.18"
-"ruamel.yaml" = {version = ">=0.16.5,<0.18.0", optional = true}
-pygls = {version = ">=0.10,<0.12", optional = true}
-dataclasses = {version = ">=0.7,~0", python = "~3.6"}
+python = ">=3.9,<4"
+transitions = ">=0.7,<0.10"
+Pint = ">=0.15,<0.23"
+"ruamel.yaml" = {version = ">=0.16.5,<0.19.0", optional = true}
+pygls = {version = "^1.0.0", optional = true}
 Jinja2 = ">=2.11.3,<4.0.0"
-pydantic = "^1.8.2"
+pydantic = ">=2,<3.0"
 click = ">=7.1.2,<9"
 
 [tool.poetry.extras]
@@ -31,15 +30,14 @@ yaml = ["ruamel.yaml"]
 lsp = ["pygls"]
 
 [tool.poetry.dev-dependencies]
-pytest = "^7.0"
-codecov = "^2.0.15"
-pytest-cov = "^3.0.0"
-pytest-console-scripts = "^1.0.0"
-sphinx = "^4.0.1"
-sphinx-rtd-theme = "^1.0.0"
-taskipy = "^1.4.0"
-mypy = "^0.942"
-sphinx-click = "^3.0.1"
+pytest = "^7.1"
+pytest-cov = "^4.0"
+pytest-console-scripts = "^1.4"
+sphinx = "^7.2.6"
+sphinx-rtd-theme = "^2.0.0"
+taskipy = "^1.4"
+mypy = "^1.2.0"
+sphinx-click = "^5.0.1"
 
 [tool.poetry.scripts]
 cp2klint = 'cp2k_input_tools.cli.lint:cp2klint'
@@ -52,11 +50,14 @@ cp2k-datafile-lint = 'cp2k_input_tools.cli.datafile_lint:cp2k_datafile_lint'
 
 [tool.black]
 line-length = 132
-target_version = ['py36']
+target_version = ['py39']
 
-[tool.isort]
-profile = "black"
-multi_line_output = 3
+[tool.ruff]
+# Enable flake8-bugbear (`B`) and isort (`I`) rules.
+select = ["E", "F", "B", "I"]
+line-length = 136
+# ignore missing strict= argument for zip for py<3.10 compat
+ignore = ["B905"]
 
 [tool.taskipy.tasks]
 test = "pytest"
