@@ -5,18 +5,19 @@ import itertools
 from decimal import Decimal, InvalidOperation
 from typing import Iterator, List, Sequence
 
-from pydantic import BaseModel, Extra, Field, root_validator
+from pydantic import BaseModel, Field, model_validator
 
-from .utils import DatafileIterMixin, FromDictMixin, dformat
+from ..utils import DatafileIterMixin, FromDictMixin, dformat
 
 
 class PseudopotentialDataLocal(BaseModel):
     r: Decimal
     coefficients: List[Decimal] = Field(..., alias="coeffs")
 
-    class Config:
-        extra = "forbid"
-        allow_population_by_field_name = True
+    model_config = {
+        "extra": "forbid",
+        "populate_by_name": True,
+    }
 
 
 class PseudopotentialDataNonLocal(BaseModel):
@@ -24,19 +25,18 @@ class PseudopotentialDataNonLocal(BaseModel):
     nproj: int
     coefficients: List[Decimal] = Field(..., alias="coeffs")
 
-    @root_validator
-    def check_coefficients(cls, values):
-        assert (
-            len(values["coefficients"]) == values["nproj"] * (values["nproj"] + 1) // 2
-        ), "invalid number of coefficients for non-local projection"
-        return values
+    @model_validator(mode="after")
+    def check_coefficients(cls, obj):
+        assert len(obj.coefficients) == obj.nproj * (obj.nproj + 1) // 2, "invalid number of coefficients for non-local projection"
+        return obj
 
-    class Config:
-        extra = "forbid"
-        allow_population_by_field_name = True
+    model_config = {
+        "extra": "forbid",
+        "populate_by_name": True,
+    }
 
 
-class PseudopotentialDataNLCC(BaseModel, extra=Extra.forbid):
+class PseudopotentialDataNLCC(BaseModel, extra="forbid"):
     """Nonlinear Core Correction data"""
 
     r: Decimal
@@ -44,7 +44,7 @@ class PseudopotentialDataNLCC(BaseModel, extra=Extra.forbid):
     c: Decimal
 
 
-class PseudopotentialData(BaseModel, DatafileIterMixin, FromDictMixin, extra=Extra.forbid):
+class PseudopotentialData(BaseModel, DatafileIterMixin, FromDictMixin, extra="forbid"):
     element: str
     identifiers: List[str]
     n_el: List[int]
@@ -165,10 +165,10 @@ class PseudopotentialData(BaseModel, DatafileIterMixin, FromDictMixin, extra=Ext
         if self.nlcc:
             yield f" NLCC {len(self.nlcc):{i_fmt}}"
 
-            r_nlcc_max_exp = -min(nlcc.r.as_tuple().exponent for nlcc in self.nlcc)
+            r_nlcc_max_exp = -min(int(nlcc.r.as_tuple().exponent) for nlcc in self.nlcc)
             r_nlcc_max_len = max(6 + r_nlcc_max_exp, *(len(f"{nlcc.r:.{r_nlcc_max_exp}f}") for nlcc in self.nlcc))
 
-            c_nlcc_max_exp = -min(nlcc.c.as_tuple().exponent for nlcc in self.nlcc)
+            c_nlcc_max_exp = -min(int(nlcc.c.as_tuple().exponent) for nlcc in self.nlcc)
             c_nlcc_max_len = max(6 + c_nlcc_max_exp, *(len(f"{nlcc.c:.{c_nlcc_max_exp}f}") for nlcc in self.nlcc))
 
             for nlcc in self.nlcc:
