@@ -1,6 +1,7 @@
 import pytest
 
 from cp2k_input_tools.basissets import BasisSetData
+from cp2k_input_tools.utils import SYM2NUM
 
 from . import TEST_DIR
 
@@ -46,7 +47,19 @@ def test_check_formatting():
 def test_datafile_lint(script_runner):
     """check that reformatting a formatted file leaves it as is"""
     bsetfile = INPUTS_DIR / "BASIS_SET.formatted"
-    ret = script_runner.run("cp2k-datafile-lint", "basisset", str(bsetfile))
+    ret = script_runner.run(["cp2k-datafile-lint", "basisset", str(bsetfile)])
+
+    assert ret.stderr == ""
+    assert ret.success
+    assert ret.stdout == bsetfile.read_text()
+
+
+def test_datafile_lint_crystal(script_runner):
+    """check that reformatting a formatted file leaves it as is"""
+    bsetfile = INPUTS_DIR / "29_Cu.pob-DZVP-rev2"
+    ret = script_runner.run(
+        ["cp2k-datafile-lint", "basisset", "--input-basis-format", "crystal", "--output-basis-format", "crystal", str(bsetfile)]
+    )
 
     assert ret.stderr == ""
     assert ret.success
@@ -227,5 +240,14 @@ def test_bset_from_dicts():
     }
 
     # NOTE: they are not identical since the first one goes via the internal bit-representation of the float
-    BasisSetData.parse_obj(floated_dict)
-    BasisSetData.parse_obj(stringified_dict)
+    BasisSetData.model_validate(floated_dict)
+    BasisSetData.model_validate(stringified_dict)
+
+
+def test_new_style_ae_basisset_import():
+    """Test that for new style -ae basis sets the number of electrons is properly set"""
+
+    with (TEST_DIR / "inputs" / "BASIS_MOLOPT.new_style_ae").open() as fhandle:
+        for entry in BasisSetData.datafile_iter(fhandle):
+            assert entry.n_el, "Number of electrons is not set"
+            assert entry.n_el == SYM2NUM[entry.element], "Invalid number of electrons"
