@@ -1,5 +1,5 @@
 import pathlib
 
-__version__ = "0.8.2"
+__version__ = "0.9.0"
 
 DEFAULT_CP2K_INPUT_XML = pathlib.Path(__file__).resolve().parent.joinpath("cp2k_input.xml")
