@@ -3,6 +3,7 @@ import pathlib
 import click
 
 from cp2k_input_tools.basissets import BasisSetData
+from cp2k_input_tools.basissets.crystal import BasisSetData as BasisSetDataCrystal
 from cp2k_input_tools.pseudopotentials import PseudopotentialData
 
 from . import smart_open
@@ -18,7 +19,25 @@ from . import smart_open
     "fpath", metavar="[<file to lint>]", type=click.Path(dir_okay=False, allow_dash=True, path_type=pathlib.Path), default="-"
 )
 @click.option("-i", "--inplace", is_flag=True, help="replace the original file with the linted/prettified one")
-def cp2k_datafile_lint(iformat, fpath, inplace):
+@click.option(
+    "--input-basis-format",
+    type=click.Choice(("cp2k", "crystal")),
+    default="cp2k",
+    help="format in which the input basis set is specified",
+)
+@click.option(
+    "--output-basis-format",
+    type=click.Choice(("cp2k", "crystal", "nwchem-ecp")),
+    default="cp2k",
+    help="format in which the basis set should be printed",
+)
+@click.option(
+    "--identifier",
+    type=str,
+    default="",
+    help="override the identifier on output",
+)
+def cp2k_datafile_lint(iformat, fpath, inplace, input_basis_format, output_basis_format, identifier):
     """Linter/Pretty-printer for other CP2K data formats
 
     Supported formats are:
@@ -28,7 +47,15 @@ def cp2k_datafile_lint(iformat, fpath, inplace):
     """
 
     if iformat in ("basis", "basisset", "basissets"):
-        datafile = BasisSetData
+        if input_basis_format == "cp2k":
+            datafile = BasisSetData
+            if output_basis_format != "cp2k":
+                raise click.UsageError("Basis set format conversion to CRYSTAL is not yet supported")
+        elif input_basis_format == "crystal":
+            datafile = BasisSetDataCrystal
+        else:
+            raise click.UsageError("Invalid input basis format")
+
     elif iformat in ("pseudo", "pseudos", "pseudopotential", "pseudopotentials", "potentials"):
         datafile = PseudopotentialData
 
@@ -46,10 +73,24 @@ def cp2k_datafile_lint(iformat, fpath, inplace):
                     print(entry, file=fouthandle)
                     has_preceeding_comments = True
                 else:
-                    if not has_preceeding_comments:  # if there is no comment before the next entry, add at least an empty one
+                    if (
+                        not has_preceeding_comments and output_basis_format == "cp2k"
+                    ):  # if there is no comment before the next entry, add at least an empty one
                         print("#", file=fouthandle)
 
-                    for line in entry.cp2k_format_line_iter():
+                    if output_basis_format == "cp2k":
+                        if input_basis_format == "crystal":
+                            line_iter = entry.cp2k_format_line_iter(identifier)
+                        else:
+                            line_iter = entry.cp2k_format_line_iter()
+                    elif output_basis_format == "crystal":
+                        line_iter = entry.crystal_format_line_iter()
+                    elif output_basis_format == "nwchem-ecp":
+                        line_iter = entry.nwchem_ecp_format_line_iter()
+                    else:
+                        continue
+
+                    for line in line_iter:
                         print(line, file=fouthandle)
 
                     has_preceeding_comments = False
