@@ -21,122 +21,129 @@ from typing import (
 
 from typing_extensions import Protocol
 
-SYM2NUM = {
-    "H": 1,
-    "He": 2,
-    "Li": 3,
-    "Be": 4,
-    "B": 5,
-    "C": 6,
-    "N": 7,
-    "O": 8,
-    "F": 9,
-    "Ne": 10,
-    "Na": 11,
-    "Mg": 12,
-    "Al": 13,
-    "Si": 14,
-    "P": 15,
-    "S": 16,
-    "Cl": 17,
-    "Ar": 18,
-    "K": 19,
-    "Ca": 20,
-    "Sc": 21,
-    "Ti": 22,
-    "V": 23,
-    "Cr": 24,
-    "Mn": 25,
-    "Fe": 26,
-    "Co": 27,
-    "Ni": 28,
-    "Cu": 29,
-    "Zn": 30,
-    "Ga": 31,
-    "Ge": 32,
-    "As": 33,
-    "Se": 34,
-    "Br": 35,
-    "Kr": 36,
-    "Rb": 37,
-    "Sr": 38,
-    "Y": 39,
-    "Zr": 40,
-    "Nb": 41,
-    "Mo": 42,
-    "Tc": 43,
-    "Ru": 44,
-    "Rh": 45,
-    "Pd": 46,
-    "Ag": 47,
-    "Cd": 48,
-    "In": 49,
-    "Sn": 50,
-    "Sb": 51,
-    "Te": 52,
-    "I": 53,
-    "Xe": 54,
-    "Cs": 55,
-    "Ba": 56,
-    "La": 57,
-    "Ce": 58,
-    "Pr": 59,
-    "Nd": 60,
-    "Pm": 61,
-    "Sm": 62,
-    "Eu": 63,
-    "Gd": 64,
-    "Tb": 65,
-    "Dy": 66,
-    "Ho": 67,
-    "Er": 68,
-    "Tm": 69,
-    "Yb": 70,
-    "Lu": 71,
-    "Hf": 72,
-    "Ta": 73,
-    "W": 74,
-    "Re": 75,
-    "Os": 76,
-    "Ir": 77,
-    "Pt": 78,
-    "Au": 79,
-    "Hg": 80,
-    "Tl": 81,
-    "Pb": 82,
-    "Bi": 83,
-    "Po": 84,
-    "At": 85,
-    "Rn": 86,
-    "Fr": 87,
-    "Ra": 88,
-    "Ac": 89,
-    "Th": 90,
-    "Pa": 91,
-    "U": 92,
-    "Np": 93,
-    "Pu": 94,
-    "Am": 95,
-    "Cm": 96,
-    "Bk": 97,
-    "Cf": 98,
-    "Es": 99,
-    "Fm": 100,
-    "Md": 101,
-    "No": 102,
-    "Lr": 103,
-    "Rf": 104,
-    "Db": 105,
-    "Sg": 106,
-    "Bh": 107,
-    "Hs": 108,
-    "Mt": 109,
-    "Ds": 110,
-    "Rg": 111,
-    "Cn": 112,
-    "Fl": 114,
-    "Lv": 116,
-}
+NUM2SYM = [
+    "",
+    "H",
+    "He",
+    "Li",
+    "Be",
+    "B",
+    "C",
+    "N",
+    "O",
+    "F",
+    "Ne",
+    "Na",
+    "Mg",
+    "Al",
+    "Si",
+    "P",
+    "S",
+    "Cl",
+    "Ar",
+    "K",
+    "Ca",
+    "Sc",
+    "Ti",
+    "V",
+    "Cr",
+    "Mn",
+    "Fe",
+    "Co",
+    "Ni",
+    "Cu",
+    "Zn",
+    "Ga",
+    "Ge",
+    "As",
+    "Se",
+    "Br",
+    "Kr",
+    "Rb",
+    "Sr",
+    "Y",
+    "Zr",
+    "Nb",
+    "Mo",
+    "Tc",
+    "Ru",
+    "Rh",
+    "Pd",
+    "Ag",
+    "Cd",
+    "In",
+    "Sn",
+    "Sb",
+    "Te",
+    "I",
+    "Xe",
+    "Cs",
+    "Ba",
+    "La",
+    "Ce",
+    "Pr",
+    "Nd",
+    "Pm",
+    "Sm",
+    "Eu",
+    "Gd",
+    "Tb",
+    "Dy",
+    "Ho",
+    "Er",
+    "Tm",
+    "Yb",
+    "Lu",
+    "Hf",
+    "Ta",
+    "W",
+    "Re",
+    "Os",
+    "Ir",
+    "Pt",
+    "Au",
+    "Hg",
+    "Tl",
+    "Pb",
+    "Bi",
+    "Po",
+    "At",
+    "Rn",
+    "Fr",
+    "Ra",
+    "Ac",
+    "Th",
+    "Pa",
+    "U",
+    "Np",
+    "Pu",
+    "Am",
+    "Cm",
+    "Bk",
+    "Cf",
+    "Es",
+    "Fm",
+    "Md",
+    "No",
+    "Lr",
+    "Rf",
+    "Db",
+    "Sg",
+    "Bh",
+    "Hs",
+    "Mt",
+    "Ds",
+    "Rg",
+    "Cn",
+    "Nh",
+    "Fl",
+    "Mc",
+    "Lv",
+    "Ts",
+    "Og",
+]
+
+SYM2NUM = {sym: Z for Z, sym in enumerate(NUM2SYM[1:], start=1)}
 
 EOF_MARKER_LINE = "Eof marker"
 EMPTY_LINE_MATCH = re.compile(r"^(\s*|\s*#.*)$")
@@ -160,6 +167,10 @@ _T = TypeVar("_T")
 class SupportsFromLines(Protocol):
     @classmethod
     def from_lines(cls: Type[_T], lines: Sequence[str]) -> _T:
+        pass
+
+    @staticmethod
+    def is_block_start(line: str) -> bool:
         pass
 
 
@@ -196,16 +207,14 @@ class DatafileIterMixin:
                     yield (line.strip())
                 continue  # ignore empty and comment lines
 
-            match = BLOCK_MATCH.match(line)
-
-            if match and line_buffer:
+            if (cls.is_block_start(line) or line == EOF_MARKER_LINE) and line_buffer:
                 try:
                     yield cls.from_lines(line_buffer)
                 except ValueError as exc:
-                    newexc = chained_exception(ValueError, f"failed to parse block for: {line_buffer[0].strip()}", exc)
+                    msg = f"failed to parse block for: {line_buffer[0].strip()}"
                     if not keep_going:
-                        raise newexc
-                    errors.append(newexc)
+                        raise ValueError(msg) from exc
+                    errors.append(chained_exception(ValueError, msg, exc))
 
                 if post_comments:
                     if emit_comments:
@@ -227,14 +236,18 @@ class DatafileIterMixin:
         if errors:
             raise MulitpleValueErrorsException(errors)
 
+    @staticmethod
+    def is_block_start(line: str) -> bool:
+        return BLOCK_MATCH.match(line) is not None
+
 
 class FromDictMixin:
     @classmethod
     def from_dict(cls: Type[_T], data: Mapping[str, Any], type_hooks: Optional[Mapping[Type, Callable[[Any], Any]]] = None) -> _T:
         """Create a data instance from a nested dictionary"""
-        warnings.warn("This helper function will be removed, use '.parse_obj' instead.", PendingDeprecationWarning)
+        warnings.warn("This helper function will be removed, use '.parse_obj' instead.", PendingDeprecationWarning, stacklevel=1)
         if type_hooks:
-            warnings.warn("The 'type_hooks' attribute has been removed and is being ignored.", DeprecationWarning)
+            warnings.warn("The 'type_hooks' attribute has been removed and is being ignored.", DeprecationWarning, stacklevel=1)
 
         return cls.parse_obj(data)  # type: ignore
 
