@@ -10,7 +10,7 @@ INPUTS_DIR = TEST_DIR / "inputs"
 
 
 def test_fromcp2k_json_simple(script_runner):
-    ret = script_runner.run("fromcp2k", str(INPUTS_DIR / "test01.inp"))
+    ret = script_runner.run(["fromcp2k", str(INPUTS_DIR / "test01.inp")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -26,7 +26,7 @@ def test_fromcp2k_json_simple(script_runner):
 
 
 def test_fromcp2k_json_canonical(script_runner):
-    ret = script_runner.run("fromcp2k", "-c", str(INPUTS_DIR / "test01.inp"))
+    ret = script_runner.run(["fromcp2k", "-c", str(INPUTS_DIR / "test01.inp")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -38,7 +38,7 @@ def test_fromcp2k_json_canonical(script_runner):
 
 
 def test_fromcp2k_json_trafo_lower(script_runner):
-    ret = script_runner.run("fromcp2k", "--trafo=lower", str(INPUTS_DIR / "test01.inp"))
+    ret = script_runner.run(["fromcp2k", "--trafo=lower", str(INPUTS_DIR / "test01.inp")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -49,7 +49,7 @@ def test_fromcp2k_json_trafo_lower(script_runner):
 
 
 def test_fromcp2k_json_trafo_upper(script_runner):
-    ret = script_runner.run("fromcp2k", "--trafo=upper", str(INPUTS_DIR / "test01.inp"))
+    ret = script_runner.run(["fromcp2k", "--trafo=upper", str(INPUTS_DIR / "test01.inp")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -63,7 +63,7 @@ def test_fromcp2k_yaml_simple(script_runner):
     ryaml = pytest.importorskip("ruamel.yaml")
     yaml = ryaml.YAML()
 
-    ret = script_runner.run("fromcp2k", "-f", "yaml", str(INPUTS_DIR / "test01.inp"))
+    ret = script_runner.run(["fromcp2k", "-f", "yaml", str(INPUTS_DIR / "test01.inp")])
     assert ret.success
     assert ret.stderr == ""
 
@@ -78,7 +78,7 @@ def test_fromcp2k_yaml_simple(script_runner):
 
 
 def test_fromcp2k_with_external_var(script_runner):
-    ret = script_runner.run("fromcp2k", "-E", "HP=1", str(INPUTS_DIR / "invalid_without_set_var.inp"))
+    ret = script_runner.run(["fromcp2k", "-E", "HP=1", str(INPUTS_DIR / "invalid_without_set_var.inp")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -90,7 +90,7 @@ def test_fromcp2k_with_external_var(script_runner):
     assert "DFT" in tree["force_eval"]
     assert "kpoints" in tree["force_eval"]["DFT"]
 
-    ret = script_runner.run("fromcp2k", "-E", "HP=0", str(INPUTS_DIR / "invalid_without_set_var.inp"))
+    ret = script_runner.run(["fromcp2k", "-E", "HP=0", str(INPUTS_DIR / "invalid_without_set_var.inp")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -104,7 +104,7 @@ def test_fromcp2k_with_external_var(script_runner):
 
 
 def test_fromcp2k_aiida_calc(script_runner):
-    ret = script_runner.run("fromcp2k", "--format", "aiida-cp2k-calc", str(INPUTS_DIR / "test01.inp"))
+    ret = script_runner.run(["fromcp2k", "--format", "aiida-cp2k-calc", str(INPUTS_DIR / "test01.inp")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -114,9 +114,8 @@ def test_fromcp2k_aiida_calc(script_runner):
 
 
 def test_fromcp2k_stdin(script_runner):
-
     with (INPUTS_DIR / "test01.inp").open() as stdin:
-        ret = script_runner.run("fromcp2k", stdin=stdin)
+        ret = script_runner.run(["fromcp2k"], stdin=stdin)
 
     assert ret.success
     assert ret.stderr == ""
@@ -128,7 +127,7 @@ def test_fromcp2k_stdin(script_runner):
 
 
 def test_tocp2k_json(script_runner):
-    ret = script_runner.run("tocp2k", str(INPUTS_DIR / "test01.json"))
+    ret = script_runner.run(["tocp2k", str(INPUTS_DIR / "test01.json")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -138,7 +137,7 @@ def test_tocp2k_json(script_runner):
 
 def test_tocp2k_json_stdin(script_runner):
     with (INPUTS_DIR / "test01.json").open() as stdin:
-        ret = script_runner.run("tocp2k", stdin=stdin)
+        ret = script_runner.run(["tocp2k"], stdin=stdin)
 
     assert ret.success
     assert ret.stderr == ""
@@ -149,7 +148,7 @@ def test_tocp2k_json_stdin(script_runner):
 def test_tocp2k_yaml(script_runner):
     pytest.importorskip("ruamel.yaml")
 
-    ret = script_runner.run("tocp2k", "-y", str(INPUTS_DIR / "NaCl-BS.yaml"))
+    ret = script_runner.run(["tocp2k", "-y", str(INPUTS_DIR / "NaCl-BS.yaml")])
 
     assert ret.success
     assert ret.stderr == ""
@@ -158,14 +157,14 @@ def test_tocp2k_yaml(script_runner):
 
 
 def test_cp2klint(script_runner):
-    ret = script_runner.run("cp2klint", str(INPUTS_DIR / "test01.inp"))
+    ret = script_runner.run(["cp2klint", str(INPUTS_DIR / "test01.inp")])
 
     assert ret.success
     assert ret.stderr == ""
 
 
 def test_cp2klint_error(script_runner):
-    ret = script_runner.run("cp2klint", str(INPUTS_DIR / "error_nvar.inp"))
+    ret = script_runner.run(["cp2klint", str(INPUTS_DIR / "error_nvar.inp")])
 
     assert not ret.success
     assert "Syntax error" in ret.stdout
@@ -173,7 +172,7 @@ def test_cp2klint_error(script_runner):
 
 
 def test_cp2klint_error_context(script_runner):
-    ret = script_runner.run("cp2klint", str(INPUTS_DIR / "unterminated_var.inp"))
+    ret = script_runner.run(["cp2klint", str(INPUTS_DIR / "unterminated_var.inp")])
 
     assert not ret.success
     assert "Syntax error: unterminated variable" in ret.stdout
@@ -181,28 +180,28 @@ def test_cp2klint_error_context(script_runner):
 
 
 def test_cp2klint_invalid_set_arg(script_runner):
-    ret = script_runner.run("cp2klint", "--set", "missing-equal-sign", str(INPUTS_DIR / "test01.inp"))
+    ret = script_runner.run(["cp2klint", "--set", "missing-equal-sign", str(INPUTS_DIR / "test01.inp")])
 
     assert not ret.success
     assert "Error: Invalid value for '-E' / '--set'" in ret.stderr
 
 
 def test_cp2klint_invalid_without_set_var(script_runner):
-    ret = script_runner.run("cp2klint", str(INPUTS_DIR / "invalid_without_set_var.inp"))
+    ret = script_runner.run(["cp2klint", str(INPUTS_DIR / "invalid_without_set_var.inp")])
 
     assert not ret.success
     assert "Syntax error: undefined variable 'HP'" in ret.stdout
 
 
 def test_cp2klint_invalid_without_set_var_defined(script_runner):
-    ret = script_runner.run("cp2klint", "--set", "HP=1", str(INPUTS_DIR / "invalid_without_set_var.inp"))
+    ret = script_runner.run(["cp2klint", "--set", "HP=1", str(INPUTS_DIR / "invalid_without_set_var.inp")])
 
     assert ret.success
 
 
 def test_cp2kgen_simplified(script_runner):
     with TemporaryDirectory() as cwd:
-        ret = script_runner.run("cp2kgen", str(INPUTS_DIR / "NaCl.inp"), "force_eval/dft/mgrid/cutoff=[800,900,1000]", cwd=cwd)
+        ret = script_runner.run(["cp2kgen", str(INPUTS_DIR / "NaCl.inp"), "force_eval/dft/mgrid/cutoff=[800,900,1000]"], cwd=cwd)
 
         assert ret.stderr == ""
         assert ret.success
@@ -215,7 +214,7 @@ def test_cp2kgen_simplified(script_runner):
 def test_cp2kgen_canonical(script_runner):
     with TemporaryDirectory() as cwd:
         ret = script_runner.run(
-            "cp2kgen", "-c", str(INPUTS_DIR / "NaCl.inp"), "+force_eval/0/+dft/+mgrid/cutoff=[800,900,1000]", cwd=cwd
+            ["cp2kgen", "-c", str(INPUTS_DIR / "NaCl.inp"), "+force_eval/0/+dft/+mgrid/cutoff=[800,900,1000]"], cwd=cwd
         )
 
         assert ret.stderr == ""
@@ -228,7 +227,7 @@ def test_cp2kgen_canonical(script_runner):
 
 def test_cp2kgen_simplified_indexed_single_value(script_runner):
     with TemporaryDirectory() as cwd:
-        ret = script_runner.run("cp2kgen", str(INPUTS_DIR / "NaCl.inp"), "force_eval/subsys/cell/a/0=10.0", cwd=cwd)
+        ret = script_runner.run(["cp2kgen", str(INPUTS_DIR / "NaCl.inp"), "force_eval/subsys/cell/a/0=10.0"], cwd=cwd)
 
         assert ret.stderr == ""
         assert ret.success
@@ -239,14 +238,14 @@ def test_cp2kgen_simplified_indexed_single_value(script_runner):
 
 def test_cp2kgen_invalid_expression(script_runner):
     with TemporaryDirectory() as cwd:
-        ret = script_runner.run("cp2kgen", str(INPUTS_DIR / "NaCl.inp"), "force_eval/subsys/cell/a/0 -> test", cwd=cwd)
+        ret = script_runner.run(["cp2kgen", str(INPUTS_DIR / "NaCl.inp"), "force_eval/subsys/cell/a/0 -> test"], cwd=cwd)
 
         assert "an expression must be of the form" in ret.stderr
         assert not ret.success
 
 
 def test_cp2kget_simplified(script_runner):
-    ret = script_runner.run("cp2kget", str(INPUTS_DIR / "NaCl.inp"), "force_eval/dft/mgrid/cutoff")
+    ret = script_runner.run(["cp2kget", str(INPUTS_DIR / "NaCl.inp"), "force_eval/dft/mgrid/cutoff"])
 
     assert ret.stderr == ""
     assert ret.success
@@ -254,7 +253,7 @@ def test_cp2kget_simplified(script_runner):
 
 
 def test_cp2kget_canonical(script_runner):
-    ret = script_runner.run("cp2kget", "-c", str(INPUTS_DIR / "NaCl.inp"), "+force_eval/0/+dft/+mgrid/cutoff")
+    ret = script_runner.run(["cp2kget", "-c", str(INPUTS_DIR / "NaCl.inp"), "+force_eval/0/+dft/+mgrid/cutoff"])
 
     assert ret.stderr == ""
     assert ret.success
@@ -262,7 +261,7 @@ def test_cp2kget_canonical(script_runner):
 
 
 def test_cp2kget_simplified_indexed_single_value(script_runner):
-    ret = script_runner.run("cp2kget", str(INPUTS_DIR / "NaCl.inp"), "force_eval/subsys/cell/a/0")
+    ret = script_runner.run(["cp2kget", str(INPUTS_DIR / "NaCl.inp"), "force_eval/subsys/cell/a/0"])
 
     assert ret.stderr == ""
     assert ret.success
@@ -271,7 +270,7 @@ def test_cp2kget_simplified_indexed_single_value(script_runner):
 
 def test_cp2kget_simplified_list_value(script_runner):
     """check that getting a list element gives human readable output"""
-    ret = script_runner.run("cp2kget", str(INPUTS_DIR / "NaCl.inp"), "force_eval/subsys/cell/a")
+    ret = script_runner.run(["cp2kget", str(INPUTS_DIR / "NaCl.inp"), "force_eval/subsys/cell/a"])
 
     assert ret.stderr == ""
     assert ret.success
