@@ -14,8 +14,7 @@ if hasattr(sys, "pypy_version_info"):
 pygls = pytest.importorskip("pygls")
 
 
-from pygls.lsp.methods import TEXT_DOCUMENT_DID_OPEN  # noqa: E402
-from pygls.lsp.types import DidOpenTextDocumentParams, TextDocumentItem  # noqa: E402
+from lsprotocol.types import TEXT_DOCUMENT_DID_OPEN, DidOpenTextDocumentParams, TextDocumentItem  # noqa: E402
 
 CALL_TIMEOUT = 5
 
